Move contexts back to their own directory
