+++ src/hooks/useEditorState.ts
@@ -1,7 +1,7 @@
 import type { EditorState } from "prosemirror-state";
 import { useContext } from "react";
 
-import { EditorContext } from "../components/EditorContext.js";
+import { EditorContext } from "../contexts/EditorContext.js";
 
 /**
  * Provides access to the current EditorState value.
